Fix: admin update paasword & email
