+++ src/common.less
@@ -480,3 +480,6 @@ a {
   width: 100%;
   margin: 0 auto;
 }
+.inline-block {
+  display: inline-block;
+}
